requirements.txt
Flask
PyMySQL
python-dotenv
Werkzeug
cryptography
reportlab
Flask-APScheduler

